storage: consolidate microbenchmarks
Move microbenchmarks from bench_pebble_test.go into bench_test.go.

Epic: none
Release note: none
